Refactor telemetry analytics view into modular components
